Move shell integration refresh into debounced method
Fixes #285031
